feat: auth & authorized first step

+++ part3/hbnb/app/api/v1/reviews.py
@@ -1,5 +1,6 @@
 from flask_restx import Namespace, Resource, fields
 from app.services import facade
+from flask_jwt_extended import jwt_required, get_jwt_identity
 
 api = Namespace('reviews', description='Review operations')
 
@@ -22,10 +23,23 @@ class ReviewList(Resource):
     @api.expect(review_model)
     @api.response(201, 'Review successfully created')
     @api.response(400, 'Invalid input data')
+    @jwt_required()
     def post(self):
         """Register a new review"""
         try:
             review_data = api.payload
+            current_user = get_jwt_identity()  # Get the current user's identity from the JWT token
+            review_data['user_id'] = current_user  # Set the user_id to the current user's ID
+            place = facade.get_place(review_data['place_id'])
+            if not place:
+                return {'error': 'Place not found'}, 404
+            if place.owner.id == current_user:
+                return {'error': 'You cannot review your own place'}, 400
+            # Check that the user has not already reviewed this place.
+            review_by_id = facade.get_reviews_by_place(review_data['place_id'])
+            if any(review.user_id == current_user for review in review_by_id):
+                return {'error': 'You have already reviewed this place'}, 400
+
             new_review = facade.create_review(review_data)
             return {
                 'id': new_review.id,
@@ -72,11 +86,16 @@ class ReviewResource(Resource):
     @api.response(200, 'Review updated successfully')
     @api.response(404, 'Review not found')
     @api.response(400, 'Invalid input data')
+    @jwt_required()
     def put(self, review_id):
         """Update a review's information"""
+        current_user = get_jwt_identity()
         review = facade.get_review(review_id)
         if not review:
             return {'error': 'Review not found'}, 404
+
+        if review.user_id != current_user:
+            return {'error': 'Unauthorized action'}, 403
 
         data = api.payload
         if not data:
@@ -89,10 +108,19 @@ class ReviewResource(Resource):
 
     @api.response(200, 'Review deleted successfully')
     @api.response(404, 'Review not found')
+    @jwt_required()
     def delete(self, review_id):
         """Delete a review"""
+        current_user = get_jwt_identity()
         review = facade.get_review(review_id)
         if not review:
             return {'message': 'Review not found'}, 404
+
+        print(review.user_id)
+        print(current_user)
+        # TODO: Unauthorized delete from users
+        if review.user_id != current_user:
+            return {'error': 'Unauthorized action'}, 403
+
         facade.delete_review(review_id)
         return {'message': 'Review successfully deleted'}, 200
